Moving fe* files into subdirectory
