+++ fem/CMakeLists.txt
@@ -42,15 +42,15 @@ set(SRCS
   eltrans.cpp
   estimators.cpp
   fe.cpp
-  fe_base.cpp
-  fe_fixed_order.cpp
-  fe_h1.cpp
-  fe_l2.cpp
-  fe_nd.cpp
-  fe_nurbs.cpp
-  fe_pos.cpp
-  fe_rt.cpp
-  fe_ser.cpp
+  fe/fe_base.cpp
+  fe/fe_fixed_order.cpp
+  fe/fe_h1.cpp
+  fe/fe_l2.cpp
+  fe/fe_nd.cpp
+  fe/fe_nurbs.cpp
+  fe/fe_pos.cpp
+  fe/fe_rt.cpp
+  fe/fe_ser.cpp
   fe_coll.cpp
   fespace.cpp
   geom.cpp
@@ -130,15 +130,15 @@ set(HDRS
   eltrans.hpp
   estimators.hpp
   fe.hpp
-  fe_base.hpp
-  fe_fixed_order.hpp
-  fe_h1.hpp
-  fe_l2.hpp
-  fe_nd.hpp
-  fe_nurbs.hpp
-  fe_pos.hpp
-  fe_rt.hpp
-  fe_ser.hpp
+  fe/fe_base.hpp
+  fe/fe_fixed_order.hpp
+  fe/fe_h1.hpp
+  fe/fe_l2.hpp
+  fe/fe_nd.hpp
+  fe/fe_nurbs.hpp
+  fe/fe_pos.hpp
+  fe/fe_rt.hpp
+  fe/fe_ser.hpp
   fe_coll.hpp
   fem.hpp
   fespace.hpp
